manage Dynamic array of pickup summary

+++ src/app/containers/DashboardYearly/Cumulative/index.tsx
@@ -227,7 +227,7 @@ const CumulativeTable = ({ selectedDate }: any) => {
                                                 <div className="icon-col">
                                                   <img
                                                     src={
-                                                      parseFloat(
+                                                      parseInt(
                                                         subdata.vsBud[index]
                                                       ) < 0
                                                         ? caretdown
